Improved core and added unit tests.
Redefined how the library will be consumed to ask the user questions.
Previously, Inquisitor was a simple object and the `prompt` method took
a manifest and an array of questions to ask. Now, Inquisitor is an
object that is instantiated with a manifest (question dictionary) and
returns an object with an `ask` method that can simply be passed an
array of questions to ask.

Consequently, I adjusted the example to match the new format. I also
began adding additional unit tests to further improve testing around
this library and it's intended functionality.

There is also a shift in defining terminology for this library. See
below:

- manifest: a dictionary of questions that is used for lookups when the
  developer chooses to prompt the user.
- pivot: an object in the questions array (for `ask`) that contains a
  branch and question to pivot off of.